requirements.txt
pandas
python-dotenv
sentence-transformers
pinecone-client
langchain-text-splitters
tqdm
streamlit
google-generativeai